Change bus name "TECHNICS" => "KN5000" and move files to a subdir.
