+++ scripts/src/bus.lua
@@ -5539,15 +5539,15 @@ end
 
 ---------------------------------------------------
 --
---@src/devices/bus/technics/kn5000_extension.h,BUSES["TECHNICS"] = true
+--@src/devices/bus/technics/kn5000/kn5000_extension.h,BUSES["KN5000"] = true
 ---------------------------------------------------
 
-if (BUSES["TECHNICS"]~=null) then
+if (BUSES["KN5000"]~=null) then
 	files {
-		MAME_DIR .. "src/devices/bus/technics/kn5000_extension.cpp",
-		MAME_DIR .. "src/devices/bus/technics/kn5000_extension.h",
-		MAME_DIR .. "src/devices/bus/technics/hdae5000.cpp",
-		MAME_DIR .. "src/devices/bus/technics/hdae5000.h",
+		MAME_DIR .. "src/devices/bus/technics/kn5000/kn5000_extension.cpp",
+		MAME_DIR .. "src/devices/bus/technics/kn5000/kn5000_extension.h",
+		MAME_DIR .. "src/devices/bus/technics/kn5000/hdae5000.cpp",
+		MAME_DIR .. "src/devices/bus/technics/kn5000/hdae5000.h",
 	}
 end
