+++ src/devices/bus/technics/kn5000/hdae5000.cpp
