+++ src/devices/bus/technics/kn5000/hdae5000.h
@@ -1,8 +1,8 @@
 // license:BSD-3-Clause
 // copyright-holders:Olivier Galibert, Felipe Sanches
 
-#ifndef MAME_BUS_TECHNICS_HDAE5000_H
-#define MAME_BUS_TECHNICS_HDAE5000_H
+#ifndef MAME_BUS_TECHNICS_KN5000_HDAE5000_H
+#define MAME_BUS_TECHNICS_KN5000_HDAE5000_H
 
 // HD-AE5000 (hard-drive + parport + extra audio output)
 
@@ -12,4 +12,4 @@
 
 DECLARE_DEVICE_TYPE(HDAE5000, device_kn5000_extension_interface)
 
-#endif // MAME_BUS_TECHNICS_HDAE5000_H
+#endif // MAME_BUS_TECHNICS_KN5000_HDAE5000_H
