+++ src/devices/bus/technics/kn5000/kn5000_extension.cpp
